feat: Accommodation 엔티티에 currency 필드 추가

+++ src/main/java/kr/kro/airbob/domain/accommodation/dto/AccommodationResponse.java
@@ -45,7 +45,7 @@ public class AccommodationResponse {
 		String name,
 		String description,
 		AccommodationType type,
-		Integer basePrice,
+		Long basePrice,
 		LocalTime checkInTime,
 		LocalTime checkOutTime,
